Release 1.0

+++ motyaftg.py
@@ -1,9 +1,6 @@
-# Мотя неВеликого Хотары
-# Смесь скриптов uniborg'а и FTG
+# Адаптированная версия Моти для Friendly-Telegram
 
-# А здесь моя рука : Накодено Деменкопом
-
-
+# Оптимизирован и создан для Heroku
 
 import logging
 import os
@@ -20,7 +17,8 @@ def register(cb):
     cb(Motya())
 
 class Motya(loader.Module):
-    """Ваш лучший генератор демотиваторов, спасибо Hotaru подпишитесь на официальный канал @motyachan там есть инструкция по установке и обновления если они будут"""
+    """Простой и крутой (\U0001F60E) генератор демотиваторов.
+Чат и помощь в @motyachat"""
 
     def __init__(self):
         self.name = _("Motya")
@@ -33,23 +31,8 @@ class Motya(loader.Module):
         self._me = await client.get_me()
 
     async def motyacmd(self, message):
-        """.motya <TEXT> в ответ на картинку для создания демотиватора"""
-        # Удаление предыдущих файлов, если таковые присутствуют
-        await message.edit(_("Подготовка Моти)"))
-        try:
-            url = 'https://raw.githubusercontent.com/mishailovic/motya-cli/master/template.jpg'
-            r = requests.get(url, allow_redirects=True)
-            open('template.jpg', 'wb').write(r.content)
-        except:
-            await message.edit(_("Произошла ошибка на этапе получения темплейта"))
-
-        try:
-            url2 = 'https://github.com/mishailovic/motya-cli/blob/master/times.ttf?raw=true'
-            r2 = requests.get(url2, allow_redirects=True)
-            open('times.ttf', 'wb').write(r2.content)
-        except:
-            await message.edit(_("Шрифт сказал идти подальше"))
-
+        """Ответьте на картинку или стикер для получения демотиватора (требует текст)"""
+        # Переменные
         mot_template = 'template.jpg'
         UPPER_FONT = 'times.ttf'
         UPPER_SIZE = 45
@@ -60,46 +43,62 @@ class Motya(loader.Module):
         TEMPLATE_COORDS = (75, 45, 499, 373)
         PADDING = 10
 
-        prename = "premotya.jpg"
+        prename = "premotya.png"
         name = "motya.png"
+        # Удаление предыдущих файлов, если таковые присутствуют
+        if os.path.isfile(prename): os.remove(prename)
+        if os.path.isfile(name): os.remove(name)
+        if os.path.isfile(mot_template): os.remove(mot_template)
+        if os.path.isfile(UPPER_FONT): os.remove(UPPER_FONT)
+        # Скачивание
+        await message.edit(_("<code>Подготовка Моти)</code>"))
+        try:
+            url = 'https://raw.githubusercontent.com/mishailovic/motya-cli/master/template.jpg'
+            r = requests.get(url, allow_redirects=True)
+            open('template.jpg', 'wb').write(r.content)
+        except:
+            await message.edit(_("<code>Произошла ошибка на этапе получения темплейта</code>"))
 
-        if os.path.isfile(prename):
-            os.remove(prename)
-
-        if os.path.isfile(name):
-            os.remove(name)
+        try:
+            url2 = 'https://github.com/mishailovic/motya-cli/blob/master/times.ttf?raw=true'
+            r2 = requests.get(url2, allow_redirects=True)
+            open('times.ttf', 'wb').write(r2.content)
+        except:
+            await message.edit(_("<code>Шрифт сказал идти подальше</code>"))
         # Аргументики
-        args = str(utils.get_args(message))
+        args = utils.get_args(message)
         if len(args) < 1:
-            await message.edit(_("АГде текст?"))
+            await message.edit(_("<code>Использование: .motya {текст}</code>"))
             return
+        else:
+            args = str(args)
+            args = args.replace('[', '')
+            args = args.replace(']', '')
+            args = args.replace(',', '')
+            args = args.replace("'", "")
 
         # Сохранение файла
         img = await message.get_reply_message()
         if img and img.media:
-            await message.edit(_("Медиа объект получен. В процессе.."))
+            await message.edit(_("<code>Медиа объект получен. В процессе..</code>"))
             photo = io.BytesIO()
             await bot.download_media(img, photo)
         else:
-            await message.edit(_("Нужно писать в ответ на картинку"))
+            await message.edit(_("<code>Мотя требует ответа на смс, спасибо</code>"))
             return
 
         if photo:
-            await message.edit(_("Обработка..."))
+            await message.edit(_("<code>Обработка...</code>"))
             try:
                 image = Image.open(photo)
             except OSError:
-                await message.edit(_("Нам сказали идти в жопу, файл не поддерживается"))
+                await message.edit(_("<code>Тип файла не поддерживается, используйте фото или стикеры</code>"))
                 return
-            image.save(prename)
+            image.save(prename, "PNG")
             image.close()
-            await message.edit(_("Фото сохранено, начинается магия..."))
+            await message.edit(_("<code>Фото сохранено, мотя идет в дело...</code>"))
         # Собственно Мотя
         def draw_x_axis_centered_text(image, text, font, size, pos_y):
-            text = text.replace('[', '')
-            text = text.replace(']', '')
-            text = text.replace(',', '')
-            text = text.replace("'", "")
             draw = ImageDraw.Draw(image)
             text_font = ImageFont.truetype(font, size)
             text_width, text_baseline = text_font.getsize(text)
@@ -119,15 +118,14 @@ class Motya(loader.Module):
             frame.paste(demot, TEMPLATE_COORDS)
             draw_x_axis_centered_text(frame, args, UPPER_FONT, UPPER_SIZE, UPPER_FONT_Y)
             frame.save(name, "PNG")
-        make_image()
 
-        await message.edit(_("Мотя то закончила работу, отправка смс!"))
+        make_image()
+        await message.edit(_("<code>Отправка фотографии, ожидание</code>"))
         try:
             await img.client.send_file(entity = await img.client.get_input_entity(img.chat_id), file=name)
         except:
-            await message.edit(_("Ошибка при отправке"))
-
-
+            await message.edit(_("<code>Ошибка при отправке(</code>"))
+        await message.delete()
 
         if os.path.isfile(prename): os.remove(prename)
         if os.path.isfile(name): os.remove(name)
